[change] 分割ws和ws-auto的interfaces
