+++ src/interfaces/ws.ts
@@ -1,5 +1,5 @@
 import { Subtitle } from ".";
-import { AutoList, AutoSub } from "./autoplay";
+import { s2cAddAutoSubBody, s2cGetRoomAutoListsBody } from "./ws-auto";
 
 // c2s -> client to server msg
 
@@ -165,25 +165,6 @@ export interface c2sChangeReversed {
   };
 }
 
-export interface c2sAddAutoSub {
-  head: {
-    cmd: "addAutoSub";
-  };
-  body: {
-    auto_subs: AutoSub[];
-    memo: string;
-  };
-}
-
-export interface c2sGetRoomAutoLists {
-  head: {
-    cmd: "getRoomAutoLists";
-  };
-  body: {
-    room_id: number;
-  };
-}
-
 export interface c2sHeartBeat {
   // 目前的心跳是复读client发过去的东西, 然后发给心跳方
   head: {
@@ -220,7 +201,22 @@ export interface s2cEventMap {
       | "sGetRoomAutoLists"
       | "heartBeat"; // 因为目前心跳是复读所以这里是heartBeat,前面不带服务器的s
   };
-  body: any; // eslint-disable-line
+  body:
+    | s2cChangeUserBody
+    | s2cGetRoomSubBody
+    | s2cAddSubtitleBody
+    | s2cChangeSubtitleBody
+    | s2cEditChangeBody
+    | s2cAddTranslatedSubtitleBody
+    | s2cDeleteSubtitleBody
+    | s2cReorderSubBody
+    | s2cSendSubtitleBody
+    | s2cChangeStyleBody
+    | s2cChangeBilingualBody
+    | s2cChangeReversedBody
+    | s2cAddAutoSubBody
+    | s2cGetRoomAutoListsBody
+    | s2cHeartBeatBody;
 }
 
 export interface s2cChangeUserBody {
@@ -288,16 +284,6 @@ export interface s2cChangeBilingualBody {
 
 export interface s2cChangeReversedBody {
   reversed: boolean;
-}
-
-export interface s2cAddAutoSubBody {
-  status: boolean;
-  new_list: AutoList;
-}
-
-export interface s2cGetRoomAutoListsBody {
-  status: boolean;
-  auto_lists: AutoList[];
 }
 
 export interface s2cHeartBeatBody {
